Config handler implementation.

+++ src/main/java/com/xenojoshua/xjf/config/interfaces/XjfConfigHandlerInterface.java
@@ -26,11 +26,11 @@ public interface XjfConfigHandlerInterface {
     int loadInt(String key);
 
     /**
-     * Load float resource in config file.
+     * Load double resource in config file.
      * @param key
      * @return resource
      */
-    float loadFloat(String key);
+    double loadFloat(String key);
 
     /**
      * Load json resource in config file.
